added memory layout handling

+++ src/int/idt.c
@@ -50,5 +50,5 @@ void set_idt_entry_isr(enum idt_entry ent, enum int_desc_gate_type gatetype,
 {
         idt[ent].offsethigh = OFFSET_HIGH(isr);
         idt[ent].offsetlow = OFFSET_LOW(isr);
-        idt[ent].typeattr = (uint8_t)gatetype;
+        idt[ent].typeattr = gatetype;
 }
