+++ src/io/port.c
@@ -4,7 +4,7 @@ void write_port_byte(enum port p, uint8_t data)
 {
         __asm__("outb %0, %1\n"
                 :
-                : "a" (data), "Nd" ((uint16_t)p));
+                : "a" (data), "Nd" (p));
 }
 
 uint8_t read_port_byte(enum port p)
@@ -12,7 +12,7 @@ uint8_t read_port_byte(enum port p)
         uint8_t res;
         __asm__("inb %1, %0\n"
                 : "=a" (res)
-                : "Nd" ((uint16_t)p));
+                : "Nd" (p));
         return res;
 }
 
@@ -20,7 +20,7 @@ void write_port_word(enum port p, uint16_t data)
 {
         __asm__("outw %0, %1\n"
                 :
-                : "a" (data), "Nd" ((uint16_t)p));
+                : "a" (data), "Nd" (p));
 }
 
 uint16_t read_port_word(enum port p)
@@ -28,7 +28,7 @@ uint16_t read_port_word(enum port p)
         uint16_t res;
         __asm__("inw %1, %0\n"
                 : "=a" (res)
-                : "Nd" ((uint16_t)p));
+                : "Nd" (p));
         return res;
 }
 
@@ -36,7 +36,7 @@ void write_port_long(enum port p, uint32_t data)
 {
         __asm__("outl %0, %1\n"
                 :
-                : "a" (data), "Nd" ((uint16_t)p));
+                : "a" (data), "Nd" (p));
 }
 
 uint32_t read_port_long(enum port p)
@@ -44,7 +44,7 @@ uint32_t read_port_long(enum port p)
         uint32_t res;
         __asm__("inl %1, %0\n"
                 : "=a" (res)
-                : "Nd" ((uint16_t)p));
+                : "Nd" (p));
         return res;
 }
